Moved the classmethods to bounded partial 3D array
These methods should not be used with the unbounded variant
